moved some methods from conversion to parameters
